+++ aac/doc.go
@@ -1,4 +1,4 @@
 /*
-Package aac - parse and generate AAC meta data including ADTS headers.
+Package aac parses and generates AAC meta data including ADTS headers.
 */
 package aac