+++ avc/doc.go
@@ -1,4 +1,4 @@
 /*
-Package avc -  parse AVC(H.264) NAL unit headers, slice headers and complete SPS and PPS.
+Package avc parses AVC (H.264) NAL unit headers, slice headers and complete SPS and PPS.
 */
 package avc