+++ cmd/doc.go
@@ -1,12 +1,15 @@
 /*
-Package cmd - command line tools built using mp4ff.
+Package cmd provides command line tools built using mp4ff.
 
 Install like
 
 	go install ./...
 
-or remotely as
+or directly from the repo
+as
 
-	go get -u github.com/Eyevinn/mp4ff/cmd/mp4ff-info
+	go install github.com/Eyevinn/mp4ff/cmd/mp4ff-info
+	go install github.com/Eyevinn/mp4ff/cmd/mp4ff-subslister
+	...
 */
 package cmd