+++ cmd/mp4ff-crop/main.go
@@ -1,8 +1,3 @@
-/*
-mp4ff-crop crops a (progressive) mp4 file to just before a sync frame after specified number of milliseconds.
-The intension is that the structure of the file shall be left intact except for cropping of samples and
-moving mdat to the end of the file, if not already there.
-*/
 package main
 
 import (
@@ -20,11 +15,11 @@ const (
 	appName = "mp4ff-crop"
 )
 
-var usg = `Usage of %s:
-
-%s crops a (progressive) mp4 file to just before a sync frame after specified number of milliseconds.
+var usg = `%s crops a (progressive) mp4 file to just before a sync frame after specified number of milliseconds.
 The goal is to leave the file structure intact except for cropping of samples and
 moving mdat to the end of the file, if not already there.
+
+Usage of %s:
 `
 
 type options struct {
@@ -61,7 +56,6 @@ func run(args []string, stdout io.Writer) error {
 
 	if err != nil {
 		if errors.Is(err, flag.ErrHelp) {
-			fs.Usage()
 			return nil
 		}
 		return err