+++ cmd/mp4ff-encrypt/main.go
@@ -1,4 +1,3 @@
-// mp4ff-encrypt encrypts a fragmented mp4 file using Common Encryption using cenc or cbcs scheme.
 package main
 
 import (
@@ -16,12 +15,11 @@ const (
 	appName = "mp4ff-encrypt"
 )
 
-var usg = `Usage of %s:
-
-%s encrypts a fragmented mp4 file using Common Encryption with cenc or cbcs scheme.
+var usg = `%s encrypts a fragmented mp4 file using Common Encryption with cenc or cbcs scheme.
 A combined fragmented file with init segment and media segment(s) will be encrypted.
 For a pure media segment, an init segment with encryption information is needed.
 
+Usage of %s:
 `
 
 type options struct {
@@ -68,7 +66,6 @@ func run(args []string) error {
 
 	if err != nil {
 		if errors.Is(err, flag.ErrHelp) {
-			fs.Usage()
 			return nil
 		}
 		return err