+++ cmd/mp4ff-info/doc.go
@@ -1,0 +1,15 @@
+/*
+mp4ff-info prints the box tree of input mp4 (ISOBMFF) file.
+
+	Usage of mp4ff-info:
+
+		mp4ff-info [options] infile
+
+	options:
+
+		-l string
+			level of details, e.g. all:1 or trun:1,subs:1
+		-version
+			Get mp4ff version
+*/
+package main