+++ examples/combine-segs/doc.go
@@ -1,11 +1,10 @@
-# Multiplex example
+/*
+combine-segs provides an example of multiplexing tracks from fragmented MP4 files.
+It combines init and media segments from two different files into a single
+multitrack init or media segment. The functions
 
-This folder provides an example of multiplexing tracks from fragmented MP4 files.
-
-The functions
-
-     combineInitSegments
-     combineMediaSegments
+	combineInitSegments
+	combineMediaSegments
 
 combines the tracks from two or more different init or media segments.
 The combined tracks get unique track names as specified using parameters.
@@ -20,3 +19,5 @@ The data is written to the combined `mdat` box track by track, so that there
 is exactly one `trun` box for each track. Interleaving data into more `trun`
 boxes should be possible by writing fractions of samples from the different
 input files.
+*/
+package main