+++ examples/initcreator/doc.go
@@ -1,0 +1,5 @@
+/*
+initcreator shows how one can create init segments for video, audio, and subtitles.
+The codecs used are AVC, HEVC, AAC, AC3, EC3, WebVTT, and TTML.
+*/
+package main